--- frontend/src/pages/host/myInfo/MyInfo.styles.ts
@@ -1,22 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 32px;
-`;
-
-export const ProfileContainer = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-`;
-
-export const Name = styled.p`
-  ${({ theme }) => theme.typography.bodyLarge}
-  color: ${({ theme }) => theme.colors.gray06};
-`;
