+++ frontend/src/pages/host/spaceHomePage/SpaceHomePage.styles.ts
